+++ packages/abstract-document/src/abstract-document/__tests__/test-markdown/markdown-bold.tsx
@@ -15,7 +15,6 @@ export const testMarkdownBold: ExportTestDef = {
         type: "TextRun",
       },
     ],
-    isMarkdown: true,
     numbering: undefined,
     style: {
       alignment: undefined,