+++ packages/abstract-document/src/abstract-document/__tests__/test-markdown/markdown-empty-keepTogether.tsx
@@ -8,7 +8,6 @@ export const testEmptyKeepTogether: ExportTestDef = {
   expectedMarkdown: {
     children: [
       {
-        isMarkdown: true,
         children: [
           {
             nestedStyleNames: [],