+++ packages/abstract-document/src/abstract-document/__tests__/test-markdown/markdown-superscript.tsx
@@ -22,7 +22,6 @@ export const testMarkdownSuperscript: ExportTestDef = {
         type: "TextRun",
       },
     ],
-    isMarkdown: true,
     numbering: undefined,
     style: {
       alignment: undefined,